+++ MineBar/MineBar/Views/SummaryView.swift
@@ -26,6 +26,7 @@ struct SummaryView: View {
         }
         .padding(.horizontal, 16)
         .padding(.vertical, 12)
+        .frame(height: 80)
         .background(.ultraThinMaterial)
     }
 }