+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
-## [1.0.0] - Unreleased
+## [1.1] - 2026-04-14
+
+- You can now choose which values show up in the status bar (Hashrate, Power) via toggles in Settings
+- Status bar values are separated by | instead of -
+- Fixed popover wobble when values update
+
+## [1.0] - 2026-04-07
 
 Initial release.